Check for release notes file before updating release modal source

+++ esbuild.js
@@ -92,14 +92,17 @@ const entry_point = process.argv[2] || 'src/index.js';
 // update release_notes.md with version
 const release_notes_path = path.join(process.cwd(), 'src', 'modals', 'release_notes.js');
 const release_notes_lines = fs.readFileSync(release_notes_path, 'utf8').split('\n');
-for(let i = 0; i < release_notes_lines.length; i++) {
-  if(release_notes_lines[i].startsWith('import release_notes_md from')) {
-    release_notes_lines[i] = `import release_notes_md from '../../releases/${package_json.version}.md' with { type: 'markdown' };`;
-    break;
+const release_notes_version_file_exists = fs.existsSync(path.join(process.cwd(), 'releases', package_json.version + '.md'));
+if(release_notes_version_file_exists) {
+  for(let i = 0; i < release_notes_lines.length; i++) {
+    if(release_notes_lines[i].startsWith('import release_notes_md from')) {
+      release_notes_lines[i] = `import release_notes_md from '../../releases/${package_json.version}.md' with { type: 'markdown' };`;
+      break;
+    }
   }
+  const updated_release_notes_text = release_notes_lines.join('\n');
+  fs.writeFileSync(release_notes_path, updated_release_notes_text);
 }
-const updated_release_notes_text = release_notes_lines.join('\n');
-fs.writeFileSync(release_notes_path, updated_release_notes_text);
 
 // markdown plugin
 const markdown_plugin = {
